+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Prevent use of `__` in asset metadata keys in local mode
 - docker: add build logs to container image build exception messages
 
 ## [0.38.0](https://github.com/Substra/substra/releases/tag/0.38.0) - 2022-09-26